refactor: comment model app 분리
